+++ swx_core/models/team.py
@@ -42,7 +42,7 @@ class Team(TeamBase, table=True):
         id (uuid.UUID): Unique team identifier.
     """
 
-    __tablename__ = "team"
+    __tablename__ = "swx_team"
     __table_args__ = {"extend_existing": True}
 
     id: uuid.UUID = Field(default_factory=uuid.uuid4, primary_key=True)